fix(bewaveddolphin): serialize clipboard count as qint64 for 32-bit (#406)
copy() wrote the selection count via `stream << itemList.size()`, where
QList::size() returns qsizetype — 8 bytes on LP64 but 4 bytes on ILP32.
paste() always reads the count as a fixed quint64 (8 bytes), so on 32-bit
arches (i386, armhf) the stream desynced: paste read the 4-byte count plus
4 bytes of the first field, yielding a huge value that truncated to 0 on
the int cast. The paste loop then ran zero iterations and cells stayed 0.

Cast the count to qint64 so the on-the-wire width matches the read on every
architecture. No-op on 64-bit; fixes the failing TestBewavedDolphinGui
copy/cut/paste cases on 32-bit.

+++ App/BeWavedDolphin/BeWavedDolphin.cpp
@@ -950,7 +950,7 @@ void BewavedDolphin::copy(const QItemSelection &ranges, QDataStream &stream)
     const int firstRow = sectionFirstRow(ranges);
     const int firstCol = sectionFirstColumn(ranges);
     const auto itemList = m_signalTableView->selectionModel()->selectedIndexes();
-    stream << itemList.size();
+    stream << static_cast<qint64>(itemList.size());
 
     for (const auto &item : itemList) {
         const int data_ = m_model->index(item.row(), item.column()).data().toInt();
